refactor: rename line_sentence_chunker folder to line_sentence
