+++ src/ingestion_pipeline/chunking/strategies/line_sentence/chunker.py
@@ -13,9 +13,9 @@ from textractor.entities.bbox import BoundingBox
 from textractor.entities.line import Line
 
 from ingestion_pipeline.chunking.schemas import DocumentChunk, DocumentMetadata
-from ingestion_pipeline.chunking.strategies.line_sentence_chunker.chunk_builder import ChunkBuilder
-from ingestion_pipeline.chunking.strategies.line_sentence_chunker.config import LineSentenceChunkingConfig
-from ingestion_pipeline.chunking.strategies.line_sentence_chunker.sentence_detector import SentenceDetector
+from ingestion_pipeline.chunking.strategies.line_sentence.chunk_builder import ChunkBuilder
+from ingestion_pipeline.chunking.strategies.line_sentence.config import LineSentenceChunkingConfig
+from ingestion_pipeline.chunking.strategies.line_sentence.sentence_detector import SentenceDetector
 
 logger = logging.getLogger(__name__)
